Groups modyfying

+++ GroupMeeting/Pages/Groups/Delete.cshtml.cs
@@ -9,9 +9,11 @@ using GroupMeeting.Data;
 using GroupMeeting.Models;
 using Microsoft.AspNetCore.Identity;
 using GroupMeeting.Areas.Identity.Data;
+using System.Web.Http;
 
 namespace GroupMeeting
 {
+    [Authorize]
     public class DeleteModel : PageModel
     {
         private readonly GroupMeetingContext _context;
@@ -34,14 +36,12 @@ namespace GroupMeeting
                 return NotFound();
             }
             
-
             Group = await _context.Groups
-                .Include(a => a.Owner).FirstOrDefaultAsync(m => m.ID == id);
+                .Include(a => a.Owner).Include(x=>x.City).FirstOrDefaultAsync(m => m.ID == id);
             if (Group == null)
             {
                 return NotFound();
             }
-            Group.City = _context.Cities.FirstOrDefault(c => c.ID == Group.CityID);
             if (user == null || user.Id != Group.OwnerID)
                 return Redirect("./Details?id=" + id);
             return Page();
@@ -55,16 +55,16 @@ namespace GroupMeeting
             }
 
             Group = await _context.Groups.FindAsync(id);
-            var groupOwner = await _context.GroupOwner.FirstAsync(go => go.GroupID == Group.ID && go.OwnerID == Group.OwnerID);
-            var groupCity = await _context.GroupCity.FirstAsync(go => go.GroupID == Group.ID && go.CityID == Group.CityID);
+         //   var groupOwner = await _context.GroupOwner.FirstAsync(go => go.GroupID == Group.ID && go.OwnerID == Group.OwnerID);
+        //    var groupCity = await _context.GroupCity.FirstAsync(go => go.GroupID == Group.ID && go.CityID == Group.CityID);
 
-            if (Group != null)
+            if (Group != null && _userManager.GetUserId(HttpContext.User) == Group.OwnerID)
             {
                 _context.Groups.Remove(Group);
-                if(groupOwner != null)
-                    _context.GroupOwner.Remove(groupOwner);
-                if (groupCity != null)
-                    _context.GroupCity.Remove(groupCity);
+          //      if(groupOwner != null)
+        //            _context.GroupOwner.Remove(groupOwner);
+          //      if (groupCity != null)
+            //        _context.GroupCity.Remove(groupCity);
                 await _context.SaveChangesAsync();
             }
